Moved all redux related directories to 'redux' folder
