+++ src/redux/reducers/configReducer.jsx
